refactor hypixeldata into serverdata/networkdata
