+++ src/main/java/at/hannibal2/skyhanni/data/HypixelData.kt
@@ -1,10 +1,10 @@
 package at.hannibal2.skyhanni.data
 
-import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.api.event.HandleEvent
-import at.hannibal2.skyhanni.config.ConfigManager.Companion.gson
+import at.hannibal2.skyhanni.data.hypixel.location.HypixelLocation
+import at.hannibal2.skyhanni.data.hypixel.location.NetworkData
+import at.hannibal2.skyhanni.data.hypixel.location.ServerData
 import at.hannibal2.skyhanni.data.model.TabWidget
-import at.hannibal2.skyhanni.events.HypixelJoinEvent
 import at.hannibal2.skyhanni.events.IslandChangeEvent
 import at.hannibal2.skyhanni.events.LorenzChatEvent
 import at.hannibal2.skyhanni.events.LorenzTickEvent
@@ -13,8 +13,6 @@ import at.hannibal2.skyhanni.events.ProfileJoinEvent
 import at.hannibal2.skyhanni.events.ScoreboardUpdateEvent
 import at.hannibal2.skyhanni.events.WidgetUpdateEvent
 import at.hannibal2.skyhanni.events.minecraft.ClientDisconnectEvent
-import at.hannibal2.skyhanni.events.modapi.HypixelHelloEvent
-import at.hannibal2.skyhanni.events.modapi.HypixelLocationChangeEvent
 import at.hannibal2.skyhanni.features.bingo.BingoAPI
 import at.hannibal2.skyhanni.features.dungeon.DungeonAPI
 import at.hannibal2.skyhanni.features.rift.RiftAPI
@@ -31,14 +29,8 @@ import at.hannibal2.skyhanni.utils.StringUtils.removeColor
 import at.hannibal2.skyhanni.utils.TabListData
 import at.hannibal2.skyhanni.utils.UtilsPatterns
 import at.hannibal2.skyhanni.utils.repopatterns.RepoPattern
-import com.google.gson.JsonObject
-import io.github.moulberry.notenoughupdates.NotEnoughUpdates
-import net.hypixel.data.region.Environment
-import net.hypixel.data.type.ServerType
 import net.minecraft.client.Minecraft
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
-import kotlin.concurrent.thread
-import kotlin.jvm.optionals.getOrNull
 import kotlin.time.Duration.Companion.seconds
 
 @SkyHanniModule
@@ -47,14 +39,7 @@ object HypixelData {
     private val patternGroup = RepoPattern.group("data.hypixeldata")
 
     // TODO add regex tests
-    private val serverNameConnectionPattern by patternGroup.pattern(
-        "servername.connection",
-        "(?<prefix>.+\\.)?hypixel\\.net",
-    )
-    private val serverNameScoreboardPattern by patternGroup.pattern(
-        "servername.scoreboard",
-        "§e(?<prefix>.+\\.)?hypixel\\.net",
-    )
+
     private val islandNamePattern by patternGroup.pattern(
         "islandname",
         "(?:§.)*(Area|Dungeon): (?:§.)*(?<island>.*)",
@@ -63,10 +48,7 @@ object HypixelData {
         "serverid.scoreboard",
         "§7\\d+/\\d+/\\d+ §8(?<servertype>[mM])(?<serverid>\\S+).*",
     )
-    private val lobbyTypePattern by patternGroup.pattern(
-        "lobbytype",
-        "(?<lobbyType>.*lobby)\\d+",
-    )
+
     private val playerAmountPattern by patternGroup.pattern(
         "playeramount",
         "^\\s*(?:§.)+Players (?:§.)+\\((?<amount>\\d+)\\)\\s*$",
@@ -116,10 +98,10 @@ object HypixelData {
     private var lastLocRaw = SimpleTimeMark.farPast()
     private var hasScoreboardUpdated = false
 
-    var hypixelEnvironment: Environment? = null
-    var hypixelLocation: HypixelLocation? = null
-    val hypixelAlpha get() = hypixelEnvironment == Environment.BETA
-    val hypixelLive get() = hypixelEnvironment == Environment.PRODUCTION
+    val location: HypixelLocation? get() = ServerData.hypixelLocation
+
+    val hypixelAlpha get() = NetworkData.onAlpha()
+    val hypixelLive get() = NetworkData.onLive()
 
     var inLobby = false
     var inLimbo = false
@@ -139,14 +121,6 @@ object HypixelData {
 
     var skyBlockArea: String? = null
     var skyBlockAreaWithSymbol: String? = null
-
-    val server get() = hypixelLocation?.serverName ?: ""
-    val lobbyType
-        get() = hypixelLocation?.lobbyName?.let {
-            lobbyTypePattern.matchMatcher(it) { group("lobbyType") }
-        } ?: ""
-    val mode get() = hypixelLocation?.mode ?: ""
-    val map get() = hypixelLocation?.map ?: ""
 
     // TODO: replace with parsing the HypixelLocation server property
     private fun checkCurrentServerId() {
@@ -213,8 +187,7 @@ object HypixelData {
         }
     }
 
-    // This code is modified from NEU, and depends on NEU (or another mod) sending /locraw.
-    private val jsonBracketPattern = "^\\{.+}".toPattern()
+
 
     private var loggerIslandChange = LorenzLogger("debug/island_change")
 
@@ -231,8 +204,6 @@ object HypixelData {
 
     @HandleEvent
     fun onDisconnect(event: ClientDisconnectEvent) {
-        hypixelEnvironment = null
-        hypixelLocation = null
         skyBlock = false
         inLobby = false
         skyBlockArea = null
@@ -279,10 +250,6 @@ object HypixelData {
     // TODO rewrite everything in here
     @SubscribeEvent
     fun onTick(event: LorenzTickEvent) {
-        if (!LorenzUtils.inSkyBlock) {
-            checkNEULocraw()
-        }
-
         if (LorenzUtils.onHypixel && LorenzUtils.inSkyBlock) {
             loop@ for (line in ScoreboardData.sidebarLinesFormatted) {
                 skyblockAreaPattern.matchMatcher(line) {
@@ -296,12 +263,6 @@ object HypixelData {
             checkProfileName()
         }
 
-        if (!LorenzUtils.onHypixel) {
-            val environment = readEnvironmentFromWorld()
-            if (environment != null) {
-                onJoinHypixel(environment)
-            }
-        }
         if (!LorenzUtils.onHypixel) return
 
         if (!event.isMod(5)) return
@@ -332,22 +293,6 @@ object HypixelData {
             profileName = group("profile").lowercase()
             ProfileJoinEvent(profileName).postAndCatch()
         }
-    }
-
-    @HandleEvent
-    fun onHypixelHello(event: HypixelHelloEvent) {
-        onJoinHypixel(event.environment)
-    }
-
-    private fun onJoinHypixel(environment: Environment) {
-        hypixelEnvironment = environment
-        HypixelJoinEvent().postAndCatch()
-        SkyHanniMod.repo.displayRepoStatus(true)
-    }
-
-    @HandleEvent
-    fun onHypixelLocationChange(event: HypixelLocationChangeEvent) {
-        hypixelLocation = event.location
     }
 
     private fun checkSidebar() {
@@ -427,87 +372,5 @@ object HypixelData {
 
     /* Backup logic - mod API should do all the below generally */
 
-    // Backup to hypixel hello event
-    private fun readEnvironmentFromWorld(): Environment? {
-        if (!hasScoreboardUpdated) return null
-        val mc = Minecraft.getMinecraft()
-        val player = mc.thePlayer ?: return null
-
-        var hypixel = false
-        var hypixelAlpha = false
-
-        player.clientBrand?.let {
-            if (it.contains("hypixel", ignoreCase = true)) {
-                hypixel = true
-            }
-        }
-
-        serverNameConnectionPattern.matchMatcher(mc.currentServerData?.serverIP ?: "") {
-            hypixel = true
-            if (group("prefix") == "alpha.") {
-                hypixelAlpha = true
-            }
-        }
-
-        for (line in ScoreboardData.sidebarLinesFormatted) {
-            serverNameScoreboardPattern.matchMatcher(line) {
-                hypixel = true
-                if (group("prefix") == "alpha.") {
-                    hypixelAlpha = true
-                }
-            }
-        }
-
-        return when {
-            hypixel && hypixelAlpha -> Environment.BETA
-            hypixel -> Environment.PRODUCTION
-            else -> null
-        }
-    }
-
-    // Modified from NEU.
-    // NEU does not send locraw when not in SkyBlock.
-    // So, as requested by Hannibal, use locraw from
-    // NEU and have NEU send it.
-    // Remove this when NEU dependency is removed
-    private fun checkNEULocraw() {
-        if (LorenzUtils.onHypixel && hypixelLocation == null && lastLocRaw.passedSince() > 15.seconds) {
-            lastLocRaw = SimpleTimeMark.now()
-            thread(start = true) {
-                Thread.sleep(1000)
-                NotEnoughUpdates.INSTANCE.sendChatMessage("/locraw")
-            }
-        }
-    }
-
-    // Backup to hypixel location event
-    fun checkForLocraw(message: String) {
-        jsonBracketPattern.matchMatcher(message.removeColor()) {
-            try {
-                val obj: JsonObject = gson.fromJson(group(), JsonObject::class.java)
-                if (obj.has("server")) {
-
-                    val server = obj["server"].asString
-
-                    // note: lobbyType was removed from locraw some point;
-                    // it just returns the server type but still calls it gametype
-                    val gameType = obj["gametype"]?.asString
-                    val lobbyName = obj["lobbyname"]?.asString
-                    val mode = obj["mode"]?.asString
-                    val map = obj["map"]?.asString
-
-                    hypixelLocation = HypixelLocation(
-                        server,
-                        ServerType.valueOf(gameType).getOrNull(),
-                        lobbyName,
-                        mode,
-                        map
-                    )
-                }
-            } catch (e: Exception) {
-                ErrorManager.logErrorWithData(e, "Failed to parse locraw data")
-            }
-        }
-    }
 }
