+++ src/main/java/at/hannibal2/skyhanni/events/HypixelJoinEvent.kt
@@ -1,5 +1,6 @@
 package at.hannibal2.skyhanni.events
 
 import at.hannibal2.skyhanni.api.event.SkyHanniEvent
+import net.hypixel.data.region.Environment
 
-class HypixelJoinEvent : LorenzEvent()
+class HypixelJoinEvent(val environment: Environment) : LorenzEvent()
