+++ src/main/java/at/hannibal2/skyhanni/events/modapi/HypixelLocationChangeEvent.kt
@@ -1,6 +1,6 @@
 package at.hannibal2.skyhanni.events.modapi
 
 import at.hannibal2.skyhanni.api.event.SkyHanniEvent
-import at.hannibal2.skyhanni.data.HypixelLocation
+import at.hannibal2.skyhanni.data.hypixel.location.HypixelLocation
 
 class HypixelLocationChangeEvent(val location: HypixelLocation) : SkyHanniEvent()
